Finished a draft of the Purchase class
Created the generic purchase class. There's also documentation/comments on how I think it would fit into the program, especially in regards to how I think it would work with databases. I think this is the last class that needs to be created, and after this we all that's left for these classes is to get them modified if they don't work with databases and tkinter correctly, or if we need some things changed in the future of the project.