Started working on account section and rearranged some stuff.
The database file changed. Now it stores a password hash and a http URL in string form to represent the avatar or image link for hte user. As well as this in the PyProject.db there are some entries now. For the first 1 password is 'password', second one has password '123a', and third has '07032004'. Just saying in case people want to test out logging in and out without having to create new accounts. Main.py has loggedinUser which represents the userID of the current user using the site, and we will use this in many other places in the project to track the user. The button has logic so that if you're not logged in it takes you to the login page, else it takes you to your own account. There's now a 'classes' folder that holds all of our classes. The 'User' class in 'User.py' has a new method called getAttributeByName, which helps in creating labels on the userPage. Added userRegistration page that allows users to register new accounts. Created userLogin page that allows user to login into existing accounts. Both only have basic error handling. Created putting widgets and inforamtion for the userPage, but it isn't fully styled out yet. Next commit will have a userEditAccount, userDeleteAcount, userManageWallet, and possibly other necessary things