Major changes for user section
We finished login feature, log out feature, register feature, and logic for keeping track on who is using the site. We fixed the logic for the userButton, which changes to 'login' and takes you to the login page when no one is signed in. When you're signed in it changes to 'User' to take you to the user page. We changed userPage as it had unnecessary error handling, since the userLogin page guarentees that the userPage is given a valid id. In Main.py we also removed the logic for the userButton since it wasn't working, and we fixed it by making it default and then influencing that button during login and logout events. Our PyProject.db has some user data now, and now stores password hashes and image links to represent the user's password and avatar/profile pic. Added utilities.py as a place to store extra functions such as error handling functions or functions that will be repeatedly used over multiple files. I've added extra pages such as userDelete and userManageBalance which will be finished in the upcoming updates. NOTE: The styling on the userPage is not finished and it seems that showing the recent purchases is not finished yet.