Fixes to program
Removed openManageBalanceBtn from userEdits since it shouldn't be there. put that button in userPage where it belongs. Also fixed rounding issues with managing balance so now it always rounds your balance to the hundredth's place like it real life. Updated styling for userPage, and created a new function createProfileBtnsSection() to create a section for the buttons on the user profile or account page. Then finally added changePassword.py to make it so users can change their passwords. This will be included in the next commit