Finished the stylings for the userPage
Finished styling and structure for userPage and the recentPurchases section. We created a function to create that section. Purchases are shown like a list, and user attributes are also shown as a list. I couldn't do the grid style for user attributes because if attribute values such as a user's address was too long it wouldn't look good for the spacing.